+++ temba/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "10.2.4"
+__version__ = "10.3.8"
 
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.