+++ temba/flows/views.py
@@ -33,6 +33,7 @@ from temba.channels.models import Channel
 from temba.contacts.models import URN
 from temba.flows.models import Flow, FlowSession, FlowStart
 from temba.ivr.models import Call
+from temba.mailroom.client.types import Exclusions
 from temba.orgs.models import IntegrationType, Org
 from temba.orgs.views.base import (
     BaseDependencyDeleteModal,
@@ -1561,13 +1562,12 @@ class FlowCRUDL(SmartCRUDL):
             recipients = contact_search.get("recipients", [])
             groups, contacts = ContactSearchWidget.parse_recipients(self.request.org, recipients)
 
-            # queue the flow start to be started by mailroom
-            flow.async_start(
+            flow.start(
                 self.request.user,
                 groups=groups,
                 contacts=contacts,
                 query=contact_search["parsed_query"] if "parsed_query" in contact_search else None,
-                exclusions=contact_search.get("exclusions", {}),
+                exclude=Exclusions(**contact_search.get("exclusions", {})),
             )
             return super().form_valid(form)