+++ temba/orgs/tests/test_definitionexport.py
@@ -609,7 +609,7 @@ class DefinitionExportTest(TembaTest):
         self.assertContains(response, "Register Patient")
 
         # delete our flow, and reimport
-        confirm_appointment.release(self.admin)
+        confirm_appointment.release(self.admin, interrupt_sessions=False)
         self.org.import_app(exported, self.admin, site="https://app.rapidpro.io")
 
         # make sure we have the previously exported expiration